# backend/server/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Open Chat application",
        "title": "Open Chat API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/api/chat/{chat_uuid}/publish": {
            "post": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "Publish a chat and return its share UUID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chats"
                ],
                "summary": "Publish chat",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Chat UUID",
                        "name": "chat_uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Published chat info",
                        "schema": {
                            "$ref": "#/definitions/chats.SharedChatPublishResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid chat UUID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Chat not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to publish chat",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/chat/{chat_uuid}/unpublish": {
            "post": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "Remove public chat sharing",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chats"
                ],
                "summary": "Unpublish chat",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Chat UUID",
                        "name": "chat_uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Unpublish result",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid chat UUID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Chat not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to unpublish chat",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/interaction/{chat_share_uuid}": {
            "get": {
                "description": "Retrieve a published interaction chat by share UUID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chats"
                ],
                "summary": "Get shared interaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Shared chat UUID",
                        "name": "chat_share_uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Public interaction chat",
                        "schema": {
                            "$ref": "#/definitions/chats.PublicInteractionChat"
                        }
                    },
                    "400": {
                        "description": "Invalid shared chat UUID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Shared chat not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/interaction/{chat_share_uuid}/messages": {
            "get": {
                "description": "Retrieve paginated messages for a shared interaction chat",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "messages"
                ],
                "summary": "List shared interaction messages",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Shared chat UUID",
                        "name": "chat_share_uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 40,
                        "description": "Page size",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of shared messages",
                        "schema": {
                            "$ref": "#/definitions/chats.ListedMessagesPage"
                        }
                    },
                    "400": {
                        "description": "Invalid shared chat UUID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Shared chat not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Couldn't list shared messages",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/interaction/{chat_share_uuid}/status": {
            "get": {
                "description": "Retrieve deterministic active/finished status for a shared interaction chat",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chats"
                ],
                "summary": "Get shared interaction status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Shared chat UUID",
                        "name": "chat_share_uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Interaction status",
                        "schema": {
                            "$ref": "#/definitions/chats.InteractionStatusResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid shared chat UUID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Shared chat not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Unable to resolve interaction status",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/bots": {
            "post": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "Create an owner-scoped automated bot user with default runtime config",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bots"
                ],
                "summary": "Create bot",
                "parameters": [
                    {
                        "description": "Bot creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bots.CreateBotRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/bots.CreateBotResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Missing permission",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Bot name already exists for owner",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "chats.InteractionStatusResponse": {
            "type": "object",
            "properties": {
                "chat_uuid": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "latest_message_finished": {
                    "type": "boolean"
                },
                "latest_message_uuid": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                }
            }
        },
...
        "chats.ListedMessagesPage": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "rows": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/chats.ListedMessage"
                    }
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "chats.PublicInteractionChat": {
            "type": "object",
            "properties": {
                "chat_share_uuid": {
                    "type": "string"
                },
                "chat_type": {
                    "type": "string"
                },
                "config": {},
                "interaction_details": {
                    "type": "object",
                    "additionalProperties": true
                },
                "partner": {},
                "published_at": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
...
        "chats.SharedChatPublishResponse": {
            "type": "object",
            "properties": {
                "chat_share_uuid": {
                    "type": "string"
                },
                "chat_uuid": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "SessionAuth": {
            "description": "Session cookie obtained from login endpoint",
            "type": "apiKey",
            "name": "session_id",
            "in": "cookie"
        }
    },
...
}